// resources/js/portfolio/types/contato.d.ts
export interface Contato {
    created_at: string;
    email: string;
    email_icon: string;
    email_link: string;
    id: number;
    linkedin: string;
    linkedin_icon: string;
    linkedin_link: string;
    mensagem_inicial: string;
    updated_at: string;
    whatsapp: string;
    whatsapp_icon: string;
    whatsapp_link: string;
}